Only publish helper for popup date fields
