+++ client/src/main/java/com/vaadin/client/ui/VDateField.java
@@ -21,8 +21,6 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import com.google.gwt.dom.client.Element;
-import com.google.gwt.i18n.client.DateTimeFormat;
 import com.google.gwt.user.client.ui.FlowPanel;
 import com.google.gwt.user.client.ui.HasEnabled;
 import com.vaadin.client.ApplicationConnection;
@@ -38,9 +36,6 @@ import com.vaadin.client.DateTimeService;
  */
 public abstract class VDateField<R extends Enum<R>> extends FlowPanel
         implements Field, HasEnabled {
-
-    private static final String ISO_DATE_TIME_PATTERN = "yyyy-MM-dd'T'HH:mm:ss";
-    private static final String ISO_DATE_PATTERN = "yyyy-MM-dd";
 
     public static final String CLASSNAME = "v-datefield";
 
@@ -73,7 +68,6 @@ public abstract class VDateField<R extends Enum<R>> extends FlowPanel
         setStyleName(CLASSNAME);
         dts = new DateTimeService();
         currentResolution = resolution;
-        publishJSHelpers(getElement());
     }
 
     public R getCurrentResolution() {
@@ -261,66 +255,5 @@ public abstract class VDateField<R extends Enum<R>> extends FlowPanel
      * @return all available resolutions
      */
     protected abstract R[] doGetResolutions();
-
-    /**
-     * Publish methods/properties on the element to be used from JavaScript.
-     *
-     * @since
-     */
-    private native void publishJSHelpers(Element root)
-    /*-{
-        var self = this;
-        root.setISOValue = $entry(function (value) {
-           self.@VDateField::setISODate(*)(value);
-        });
-        root.getISOValue = $entry(function () {
-           return self.@VDateField::getISODate()();
-        });
-    }-*/;
-
-    /**
-     * Sets the value of the date field as a locale independent ISO date
-     * (yyyy-MM-dd'T'HH:mm:ss or yyyy-MM-dd depending on whether this is a date
-     * field or a date and time field).
-     *
-     * @param isoDate
-     *            the date to set in ISO8601 format, or null to clear the date
-     *            value
-     * @since
-     */
-    public void setISODate(String isoDate) {
-        if (isoDate == null) {
-            setDate(null);
-        } else {
-            Date date = getIsoFormatter().parse(isoDate);
-            setDate(date);
-        }
-    }
-
-    /**
-     * Gets the value of the date field as a locale independent ISO date
-     * (yyyy-MM-dd'T'HH:mm:ss or yyyy-MM-dd depending on whether this is a date
-     * field or a date and time field).
-     *
-     * @return the current date in ISO8601 format, or null if no date is set
-     *
-     * @since
-     */
-    public String getISODate() {
-        Date date = getDate();
-        if (date == null) {
-            return null;
-        } else {
-            return getIsoFormatter().format(date);
-        }
-    }
-
-    private DateTimeFormat getIsoFormatter() {
-        if (supportsTime()) {
-            return DateTimeFormat.getFormat(ISO_DATE_TIME_PATTERN);
-        } else {
-            return DateTimeFormat.getFormat(ISO_DATE_PATTERN);
-        }
-    }
 
 }
